Mypage Page Refactoring
