+++ front/src/component/meeting/MeetingDetail/MeetingApplyReasonModal.jsx
@@ -1,10 +1,9 @@
-import React from "react";
-
-import { GoX } from "react-icons/go";
 import TextArea from "../../textArea/TextArea";
 import { meetingApplyReasonState } from "../../../recoil/contentState";
 import Button from "../../button/Button";
 import useInputGlobal from "../../../customhook/useInputGlobal";
+
+import { GoX } from "react-icons/go";
 
 export default function MeetingApplyReasonModal({
   meetingApplyReasonModal,
@@ -57,8 +56,7 @@ export default function MeetingApplyReasonModal({
         <Button
           type="button"
           buttonText="신청하기"
-          disabled={!meetingApplyReasonInput.length > 0}
-          buttonDisabledStyle={meetingApplyReasonInput.length === 0}
+          disabled={meetingApplyReasonInput.length === 0}
           basicStyle="h-auto mt-2"
         />
       </div>
