+++ front/src/component/myPage/tab/TabCommonComponent.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Tabs from "./Tabs";
 
 export default function TabCommonComponent({ title }) {
