+++ front/src/component/typeWriter/TypeWriter.jsx
