+++ front/src/page/Root.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import Navbar from "../component/navbar/Navbar";
 import { Outlet } from "react-router-dom";
-import Footer from "../component/Footer";
+import Footer from "../component/footer/Footer";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ScrollRestoration } from "react-router-dom";
 import { useSetRecoilState } from "recoil";
